Rename superhello to fasthello
