+++ examples/src/python/example/python_distribution/hello/fasthello/hello_package/hello.py
