+++ examples/src/python/example/python_distribution/hello/fasthello/main.py
@@ -5,15 +5,9 @@
 from __future__ import (absolute_import, division, generators, nested_scopes, print_function,
                         unicode_literals, with_statement)
 
-from setuptools import setup, find_packages
-from distutils.core import Extension
+# hello_package is a python module within the fasthello python_distribution
+from hello_package import hello
 
 
-c_module = Extension(str('super_greet'), sources=[str('super_greet.c')])
-
-setup(
-  name='superhello',
-  version='1.0.0',
-  ext_modules=[c_module],
-  packages=find_packages(),
-)
+if __name__ == '__main__':
+  hello.hello()
