+++ examples/tests/python/example/python_distribution/hello/test_superhello/test_superhello.py
@@ -5,10 +5,10 @@
 from __future__ import (absolute_import, division, generators, nested_scopes, print_function,
                         unicode_literals, with_statement)
 
-# hello_package is a python module within the superhello python_distribution.
+# hello_package is a python module within the fasthello python_distribution.
 from hello_package import hello
 
 
 # Example of writing a test that depends on a python_dist target.
-def test_superhello():
+def test_fasthello():
   assert hello.hello() == "Super hello"
