+++ tests/python/pants_test/backend/python/tasks/test_python_distribution_integration.py
@@ -15,12 +15,12 @@ from pants_test.pants_run_integration_test import PantsRunIntegrationTest
 class PythonDistributionIntegrationTest(PantsRunIntegrationTest):
   # The paths to both a project containing a simple C extension (to be packaged into a
   # whl by setup.py) and an associated test to be consumed by the pants goals tested below.
-  superhello_project = 'examples/src/python/example/python_distribution/hello/superhello'
-  superhello_tests = 'examples/tests/python/example/python_distribution/hello/test_superhello'
-  superhello_install_requires = 'testprojects/src/python/python_distribution/superhello_with_install_requires'
+  fasthello_project = 'examples/src/python/example/python_distribution/hello/fasthello'
+  fasthello_tests = 'examples/tests/python/example/python_distribution/hello/test_fasthello'
+  fasthello_install_requires = 'testprojects/src/python/python_distribution/fasthello_with_install_requires'
 
   def test_pants_binary(self):
-    command=['binary', '{}:main'.format(self.superhello_project)]
+    command=['binary', '{}:main'.format(self.fasthello_project)]
     pants_run = self.run_pants(command=command)
     self.assert_success(pants_run)
     # Check that the pex was built.
@@ -33,23 +33,23 @@ class PythonDistributionIntegrationTest(PantsRunIntegrationTest):
     os.remove(pex)
 
   def test_pants_run(self):
-    command=['run', '{}:main'.format(self.superhello_project)]
+    command=['run', '{}:main'.format(self.fasthello_project)]
     pants_run = self.run_pants(command=command)
     self.assert_success(pants_run)
     # Check that text was properly printed to stdout.
     self.assertIn('Super hello', pants_run.stdout_data)
 
   def test_pants_test(self):
-    command=['test', '{}:superhello'.format(self.superhello_tests)]
+    command=['test', '{}:fasthello'.format(self.fasthello_tests)]
     pants_run = self.run_pants(command=command)
     self.assert_success(pants_run)
 
   def test_with_install_requires(self):
-    command=['run', '{}:main_with_no_conflict'.format(self.superhello_install_requires)]
+    command=['run', '{}:main_with_no_conflict'.format(self.fasthello_install_requires)]
     pants_run = self.run_pants(command=command)
     self.assert_success(pants_run)
     self.assertIn('United States', pants_run.stdout_data)
-    command=['binary', '{}:main_with_no_conflict'.format(self.superhello_install_requires)]
+    command=['binary', '{}:main_with_no_conflict'.format(self.fasthello_install_requires)]
     pants_run = self.run_pants(command=command)
     self.assert_success(pants_run)
     pex = os.path.join(get_buildroot(), 'dist', 'main_with_no_conflict.pex')
@@ -58,13 +58,13 @@ class PythonDistributionIntegrationTest(PantsRunIntegrationTest):
     os.remove(pex)
 
   def test_with_conflicting_transitive_deps(self):
-    command=['run', '{}:main_with_conflicting_dep'.format(self.superhello_install_requires)]
+    command=['run', '{}:main_with_conflicting_dep'.format(self.fasthello_install_requires)]
     pants_run = self.run_pants(command=command)
     self.assert_failure(pants_run)
     self.assertIn('pycountry', pants_run.stderr_data)
-    self.assertIn('superhello', pants_run.stderr_data)
-    command=['binary', '{}:main_with_conflicting_dep'.format(self.superhello_install_requires)]
+    self.assertIn('fasthello', pants_run.stderr_data)
+    command=['binary', '{}:main_with_conflicting_dep'.format(self.fasthello_install_requires)]
     pants_run = self.run_pants(command=command)
     self.assert_failure(pants_run)
     self.assertIn('pycountry', pants_run.stderr_data)
-    self.assertIn('superhello', pants_run.stderr_data)
+    self.assertIn('fasthello', pants_run.stderr_data)
