+++ tests/python/pants_test/projects/test_testprojects_integration.py
@@ -65,7 +65,7 @@ class TestProjectsIntegrationTest(ProjectIntegrationTest):
       'testprojects/tests/java/org/pantsbuild/testproject/depman:old-tests',
       'testprojects/tests/java/org/pantsbuild/testproject/htmlreport:htmlreport',
       'testprojects/tests/java/org/pantsbuild/testproject/parallel.*',
-      'testprojects/src/python/python_distribution/superhello_with_install_requires.*'
+      'testprojects/src/python/python_distribution/fasthello_with_install_requires.*'
     ]
 
     # May not succeed without java8 installed
